Manager now handles hierarchy
New class handles collapsing/expanding its piece and the data while the
manager handles collapsing itself (if hierarchy) and updating the
sections locations when one expands/collapses.

I changed the arrow which is from github repo based on iOS images. It
was black and I changed the hue to gray.

I made a custom tableview cell subclass to handle defaults on the way a
header and toggle check like cell should look. It contains an enum for
cell types besides headers and checkmark cells.