Changes to literals caused bug
When I changed to literals it caused errors since I didn’t do the right
hierarchy for the survey or the labels. I made survey too high (array
of array of array!)

I also changed labels to a single array to test both cases for a non
hierarchy. Interactions uses a double array with one array as first
argument while Labels is a flat array. Both work with the manager for a
non hierarchy.