Modal changes for UITableView (Checklist) kind
I chose the delegation route for this piece for these reasons: There is
a one to one relationship with modal to filterviewcontroller since only
one modal can be put up at one time. The rest of the application
doesn’t need to know these interactions and so using the default center
might be slight overkill to pass messages back and forth.

I put the tableview delegation code in MHCollapsibleSection since it
would be one degree away (section->filter->data) from the data compared
to the manager, which would be (manager->section->filter->data). The
section piece also didn’t have as much code as the manager, once it’s
done with all the delegations it’ll probably rival it.

MHFilterLabel now handles its own results NSDictionary and selected
results NSMutableDictionary. The user is seeing and interacting with
the NSMutableDictionary while the static one is the snapshot of the
data before they change it.

The modal the filterviewcontroller creates has a save, cancel and clear
button. When clicked, these buttons call methods to the section which
keeps track of the current modal index (current row that caused modal
to be created) so then that label can perform the updates or ignore the
updates, etc.

@michaelharro From your previous update I know you won’t be able to
review this in depth. If you get any chance to glance over it feel
free. I don’t think I can stray too far from the idea since we already
discussed what needs to be done. I’m sure some things can be done a
better way but even before integrating with MH that can be changed. I
can ask around for others who know iOS to review (at least
pre-integration) since it might be able to be used in other apps.
Thanks for your feedback so far!

I have two comments on it: The survey answers are not sorted since
they’re in a NSDictionary. I don’t know if that is a requirement or how
they would even be sorted (alphabetically??)
For the cell with a modal, I chose the disclosure indicator to indicate
that it can be tapped. I don’t know if there’s a nicer way to do this,
the disclosure button looked more helpful before ios7’s weird i that
replaced the arrow.