Fixed Bug with multiple modals
I wasn’t updating the current modal index so when I created another set
for Survey Question 2, the same data was popping up, just with a
different title. Once the change was made in the manager to call the
section to update the current modal index the data was shown/kept
properly.