Switched from Dictionary to three Arrays
When I thought about it more and looked at the android app/iPhone MH
app sorting is much easier for a user to look at an answer. MissionHub
site itself does not sort Survey Questions, they’re in quite a random
order (maybe an id). I’m not sure how the answers will be sorted before
it gets to the manager but now it’ll be up to the filterviewcontroller
to decide that.

Now there’s three arrays: one for keys, one for static values and a
mutable array for what the user changes (checking, unchecking). It was
a pretty quick change and doesn’t seem to be running slower so I think
it’ll work to get around making a tiny simple class to have a key/value
pair.