Moved properties and methods
I’m trying to find what a subclass can know about its super class. I’m
assuming the properties and methods need to be public for them in order
to update variables. After moving these methods I’m going to see if a
subclass should have access to what it does and if instead I should
encase a lot of the variables for the filterviewcontroller.