Save button and removing methods
createManagersAndPopulateData did not need to go in the super class but
rather just have a frame for it so the sub class can call it after
calling the super classes view did load
Also noticed the save button didn’t dismiss the modal, but this
actually would be better by design so the sub class doesn’t have to
override the buttonTapped method. It can just dismiss the modal itself
after dealing with the filter and notification calls to API, etc.