Fixed string file for languages
So the code was assuming the following before: # selected labels..etc.
I just learned from Alex that in Spanish it would be swapped, so #
labels selected. The code wouldn’t allow that so now I made the
identifier include the word selected for both the manager and the
sections. This way the Spanish version or any other version that would
swap these words can simply swap it in the file.