Hopefully last fix for localized strings
Some languages it might be best to stick the count after, between, I
have no idea. Anyway, now the file accounts for that and the code. I
included in the string the %i for the count and now future languages
can just include it in the count and see # selected or whatever order
we would want in the future.