Fixed bug with single label in Survey
I noticed 1 selected question for Engel’s scale by default, this is
because nothing was selected for Engel’s scale (it only has one
question) but it got the item count and added it with the getIdentifier.
So instead of 1 item, it would say 1 selected question since the
itemTitle never got overwritten with the default.