Only clear when label selected
After testing some more, I noticed that if the clear button was hit
first and then I went to a textfield type (never selected before), the
value would not save. This is because only when the row has been
selected the result arrays for textfield get initialized. The resultKey
was being set to empty string and then the check to initialize arrays
was never true so it never initialized the results or mutable results
array and then never actually saved.

This is fixed by just clearing labels that actually have changes rather
than clearing every single label whether it had changes or not. This
fixes the bug for the text fields and makes sense from a code
standpoint.