Added return code for text field
I changed the type for the text field to be a real modal when I saw how
the keyboard covered up the half modal. This does not happen on the
simulator.

I also added textFieldShouldReturn to make sure the text field resigns
properly. I cannot test this without a phone so hopefully it works
whenever I can. I will need to reformat more code and probably take
away the half modal option entirely because it doesn’t work well on the
device.