Move creation of MHCollapsibleViewManager to FilterViewController
This is some pretty significant refactoring, so I'll try to explain what I'm suggesting.  It looked like there was a lot
of object creation and manipulation in our example controllers that seemed cumbersome from a client perspective.  I noticed
that when I started to work on this with my example that

1. I was writing a lot of code that you had already written in your example, and
2. I had to have pretty extensive knowledge of the inner workings of this library.

When we extend this class, it would be nice to just create the arrays that have the labels and pass them to the super class where
all the common logic can be executed.  It would save us from having to re-implement the logic for creating ViewManagers every time
we use this class.

I anticipated having to create several functions to handle the single/double arrays, but it turned out that the same one worked for both
my very simple example and your more complex example.